Extract solicitation flag precomputation logic into a concern
