+++ spec/models/solicitation_spec.rb
@@ -515,110 +515,6 @@ end
     end
   end
 
-  describe "recent_matched_solicitations" do
-    let(:landing_subject) { create :landing_subject }
-    let(:siret) { '13000601800019' }
-    let(:email) { 'hubertine@example.com' }
-
-    let!(:parent_siret_solicitation) do
-      create :solicitation,
-             siret: siret,
-             landing_subject: landing_subject,
-             created_at: 2.weeks.ago,
-             status: 'processed'
-    end
-
-    let!(:parent_email_solicitation) do
-      create :solicitation,
-             email: email,
-             landing_subject: landing_subject,
-             created_at: 2.weeks.ago,
-             status: 'processed'
-    end
-
-    let!(:other_siret_solicitation) do
-      create :solicitation,
-             siret: '98765432100099',
-             landing_subject: landing_subject,
-             created_at: 2.weeks.ago,
-             status: 'processed'
-    end
-
-    let!(:too_old_solicitation) do
-      create :solicitation,
-             email: email,
-             siret: siret,
-             landing_subject: landing_subject,
-             created_at: 6.weeks.ago,
-             status: 'processed'
-    end
-
-    let!(:other_subject_solicitation) do
-      create :solicitation,
-             email: email,
-             siret: siret,
-             landing_subject: create(:landing_subject),
-             created_at: 2.weeks.ago,
-             status: 'processed'
-    end
-
-    let!(:no_match_solicitation) do
-      create :solicitation,
-             email: email,
-             siret: siret,
-             landing_subject: landing_subject,
-             created_at: 2.weeks.ago
-    end
-
-    let!(:child_solicitation) do
-      create :solicitation,
-             siret: siret,
-             email: email,
-             landing_subject: landing_subject
-    end
-
-    it 'displays only parent_solicitations' do
-      expect(child_solicitation.recent_matched_solicitations).to contain_exactly(parent_siret_solicitation, parent_email_solicitation)
-    end
-  end
-
-  describe "doublon_solicitations" do
-    let(:siret) { '13000601800019' }
-    let(:email) { 'hubertine@example.com' }
-
-    let!(:same_siret_solicitation) do
-      create :solicitation,
-             siret: siret
-    end
-
-    let!(:same_email_solicitation) do
-      create :solicitation,
-             email: email
-    end
-
-    let!(:same_siret_with_matched_solicitation) do
-      create :solicitation,
-             siret: siret,
-             status: :processed,
-             diagnosis: create(:diagnosis_completed)
-    end
-
-    let!(:other_siret_solicitation) do
-      create :solicitation,
-             siret: '98765432100099'
-    end
-
-    let!(:solicitation) do
-      create :solicitation,
-             siret: siret,
-             email: email
-    end
-
-    it 'displays only doublon solicitations' do
-      expect(solicitation.doublon_solicitations).to contain_exactly(same_siret_solicitation, same_email_solicitation)
-    end
-  end
-
   describe "from_intermediary?" do
     let!(:solicitation) { create :solicitation }
 
@@ -649,52 +545,6 @@ end
 
     context 'without facility' do
       it { is_expected.to be false }
-    end
-  end
-
-  describe "similar_abandonned_solicitations" do
-    let(:landing_subject) { create :landing_subject, subject: sol_subject }
-    let(:solicitation) { create :solicitation, email: 'hedy@lamarr.bzh', siret: '41816609600069' }
-    let!(:old_solicitation) { create :solicitation, email: email, siret: siret, status: status }
-    let(:has_similar_abandonned_solicitations) { solicitation.has_similar_abandonned_solicitations? }
-
-    subject { solicitation.similar_abandonned_solicitations }
-
-    context 'same email abandonned' do
-      let(:email) { 'hedy@lamarr.bzh' }
-      let(:siret) { '71816609600054' }
-      let(:status) { :canceled }
-
-      context '1 solicitation' do
-        it { is_expected.to contain_exactly(old_solicitation) }
-        it { expect(has_similar_abandonned_solicitations).to be false }
-      end
-
-      context 'many solicitations' do
-        before do
-          3.times { create :solicitation, email: email, siret: siret, status: status }
-        end
-
-        it { expect(has_similar_abandonned_solicitations).to be true }
-      end
-    end
-
-    context 'same siret abandonned' do
-      let(:email) { 'other@mail.com' }
-      let(:siret) { '41816609600069' }
-      let(:status) { :canceled }
-
-      it { is_expected.to contain_exactly(old_solicitation) }
-      it { expect(has_similar_abandonned_solicitations).to be false }
-    end
-
-    context 'same siret not abandonned' do
-      let(:email) { 'hedy@lamarr.bzh' }
-      let(:siret) { '41816609600069' }
-      let(:status) { :processed }
-
-      it { is_expected.to be_empty }
-      it { expect(has_similar_abandonned_solicitations).to be false }
     end
   end
